OCL base example

+++ ccmtools/test/ocl/Hello/run.sh
@@ -8,10 +8,10 @@ PKG=testOCL_$NAME
 OUT=test_$NAME
 VERSION=0.1
 #
-ccmtools-c++-generate -a -d -c $VERSION -p $PKG -s $OUT $IDL
-ccmtools-generate c++dbc -o $OUT $IDL
+if ! ccmtools-c++-generate -a -d -c $VERSION -p $PKG -s $OUT $IDL; then exit 1; fi
+if ! ccmtools-generate c++dbc -o $OUT $IDL; then exit 1; fi
 rm -f mdr.*
-ccmtools-c++-configure -p $PKG -s $OUT
-ccmtools-c++-make -p $PKG -s $OUT
+if ! ccmtools-c++-configure -p $PKG -s $OUT; then exit 1; fi
+if ! ccmtools-c++-make -p $PKG -s $OUT; then exit 1; fi
 echo "ccmtools-c++-make -p $PKG -s $OUT" > compile.sh
 chmod 755 compile.sh
